Construct folds (prototype, initial)
The initial idea for constructing the cross-validation folds is to
ensure that each user has a representative set of ratings that
contribute to his preferences.

Later on, we would like to figure out the behavior of the RE under
specialized situations, such as bootstrap settings when a user just
joins or a new movie has come out, but the emphasis now is to
understand baselines.