Formulate two trivial recommenders
We will use these as a baseline for comparing any non-trivial
recommenders.

The trivial recommenders simply recommend movies based on mean values
taken over the entire group.

While this keeps the complexity of the code simple (for now), the mean
values currently include the user's own rating of the movie.

We will account for this (for now) by limiting recommendations to movies
having > n ratings, n tbd. The individual contribution of the user will
be small relative to the population.

Second, trivial recommenders will serve as a baseline for recommenders
that use this simplification, so as to normalize this effect.